+++ src/player.ts
@@ -11,6 +11,7 @@ import { AdItem } from './advertising';
 import { BufferApi } from './bufferApi';
 import { VideoQuality } from './media';
 import { Network } from './network';
+import { DecoderConfigBridge } from './decoder';
 
 const PlayerModule = NativeModules.PlayerModule;
 
@@ -25,6 +26,8 @@ const PlayerModule = NativeModules.PlayerModule;
  */
 export class Player extends NativeInstance<PlayerConfig> {
   private network?: Network;
+
+  private decoderConfig?: DecoderConfigBridge;
   /**
    * Currently active source, or `null` if none is active.
    */
@@ -57,12 +60,14 @@ export class Player extends NativeInstance<PlayerConfig> {
         this.network = new Network(this.config.networkConfig);
         this.network.initialize();
       }
+      this.initDecoderConfig();
       const analyticsConfig = this.config?.analyticsConfig;
       if (analyticsConfig) {
         PlayerModule.initWithAnalyticsConfig(
           this.nativeId,
           this.config,
           this.network?.nativeId,
+          this.decoderConfig?.nativeId,
           analyticsConfig
         );
         this.analytics = new AnalyticsApi(this.nativeId);
@@ -70,9 +75,11 @@ export class Player extends NativeInstance<PlayerConfig> {
         PlayerModule.initWithConfig(
           this.nativeId,
           this.config,
-          this.network?.nativeId
+          this.network?.nativeId,
+          this.decoderConfig?.nativeId
         );
       }
+
       this.isInitialized = true;
     }
   };
@@ -85,6 +92,7 @@ export class Player extends NativeInstance<PlayerConfig> {
       PlayerModule.destroy(this.nativeId);
       this.source?.destroy();
       this.network?.destroy();
+      this.decoderConfig?.destroy();
       this.isDestroyed = true;
     }
   };
@@ -506,5 +514,22 @@ export class Player extends NativeInstance<PlayerConfig> {
       return undefined;
     }
     return PlayerModule.canPlayAtPlaybackSpeed(this.nativeId, playbackSpeed);
+  };
+
+  initDecoderConfig = () => {
+    if (this.config?.playbackConfig?.decoderConfig == null) {
+      return;
+    }
+
+    if (Platform.OS === 'ios') {
+      console.warn(
+        `[Player ${this.nativeId}] config: PlaybackConfig.DecoderConfig is not available for iOS. Only Android devices.`
+      );
+      return;
+    }
+    this.decoderConfig = new DecoderConfigBridge(
+      this.config.playbackConfig.decoderConfig
+    );
+    this.decoderConfig.initialize();
   };
 }